Introduce basket/complement quantities and custom prices
This patch gives more control of the content of a membership allowing the customization of the basket quantity, basket price and distribution price. In the same manner it allows specifing custom quantity and price on each complements. The basket size, distribution and complements are also now explicitly set on the membership making it clearer.