Introduce HalfdayParticipation Invoice
This is handy for both LDC and RDV as they need to be able to invoice their members for missing halfday participations.

The invoice can be either created from a rejected halfday participation or more generaly by just giving a member and a date.

The paid missing halfday works are counted as regonized halfday participations, the whole halfday works counting as been revamp too and is now more details in the membership page.