Improve Billing::BAS payments retrieval
This patch improves the Billing::BAS payments retrieval by ensuring that each similar payment line is only group to itself instead of just the date. Using the data was causing issue as it seems that payments that were made in the past are not appearing all in the same, which cause isr_data to be different as the days past for the same payment.