Switch to passwordless member sessions
This patch moves away from the never changing short member token that was used to access the member page. This was easy and served us well but it wasn't really secure. As I'm planning to add more private info in the member page (billing) switching to a more secure passwordless is a nice improvement without adding much complexity over the previous approach for the member.