Improve Member section with new billing and absence pages
This patch adds more feature to the member page by adding a new top menu which allows to navigates to two more new sections.

The first one is about billing, showing the number of open invoices and the past payments and invoices. This should provide more transparancy around billing.

The second one is about absences, giving the possibility to the member to enter is future absence on its own and also provides a list of the past ones.