Depot deliveries configuration
This patch allows the deliveries to be configured by depot, silimarly to 
the basket_complements. 

This is required by croQu'terre that have depots that are only delivered 
every other weeks.

This change affects a lot of code parts as previously it was just 
assumed that all depots were delivered all the time.