Billing::Snapshot, rake task and dashboard link
This patch adds a new rake task to capture billing snapshots at the end
of each ACP quarter, as the Heroku scheduled job runs every 10.min
we tried to create or update the snapshot the last hour of the quarter.

New links to download the last 4 quarter snapshots is also added at the
bottom of the billing dashboard.