RecurringBilling, improce amount fraction logic
This patch ensures that membership that ends before the end of the fiscal year,
are charged with an amount fraction that is taking that into account.

Example:
Membership is charged mensually, ends in June (6) instead of December (12).
If we are in March (3), the fraction will be 3, instead of 9 previously.

This spreads the mensual invoices better over the duration of the membership,
and avoid having a larger invoice amount the last membership's month.