RecurringBilling, fix calculation when membership ended_on is in the past
This patch follows-up b1c69fbf2577f4aeb65fde241e72cdfd88e5b53f and
fix the logic when the membership ended_on is in the past.

The `membership_amount_description` method has also been fixed, to
always use the member billing year division, as the annual one was wrongly
picked for the latest fraction when on mensual division.