Member page, allow member edition
This patch adds a new `/account` and `/account/edit` paths to the member section
that lets the signed member modifying its name, address, emails, and phones easily.

To avoid any confusion or misued, each change on these attributes are now tracked,
that way modifications history could be shown in the member admin page if required.