Recurring billing week day settings
This patch adds a  new `recurring_billing_wday` ACP settings, that allows to
configure which day of the week the recurrent and automatic billing is happening.

This replaces the `recurring_billing` feature, that was happening on every
Tuesday without the possibility to configure it.

If `recurring_billing_wday` is nil, then the recurring billing is disabled.