Remove Member#newsletter? logic
This one was used to opt-in inactive members of Rage de Vert
in the newsletter but it was confusing for all other ACP.

This patch removes the MERGE tag from MailChimp, and each
ACP can create there own segments logic and leave inactive
members to unsubuscribe on their own if they aren't intersted
anymore.

Deleted members are also now unsubscribed instead of deleted
on MailChimp as it was sometimes causing issues overthere.