Start to bring back email template to the app
This patch starts to bring back all the email templates currently
hosted by Postmark to the app direclty.

This patch only migrates the new session email template, but the plan
is to migrates them all inside the app.

Some templates like that one should required to be edited by every ACP
but other templates will be directly editable, in every language, within
the app, using the Trix editor and Liquid templates.

Postmark templates are great and allow a complete customisation but in
really only the content of these emails really needs to be edited, as
every ACP seems to be happy with the default layout. The Postmark
interface isn't the best to edit these template and must be done in
both HTML and Text version and the mutli-languages template support
isn't great and require technical skill. The in app template editing
should be much simpler and nicer to use.