Add support for Swiss QR-bill
This patch introduces the support for the new Swiss QR-Code bill that
is replacing the ISR/BVR orange bill standard.

The new QR-IBAN with a reference number is supported, the reference
number is silimar to the ISR/BVR one and parsing the payments via the
bank connection should continue to work without any change required.

Based on the ACP `qr_iban` or `ccp` presence either an QR-bill or an
orange ISR/BVR bill will be appended at the end of the invoice PDF.

Swiss QR-bill specifications:
https://www.paymentstandards.ch/dam/downloads/ig-qr-bill-en.pdf