Introduce MailTemplate
This patch introduces the new MailTemplate object, that will allow each
ACP to easily preview and edit all member email templates that are
currently hosted on Postmark. The email will still be delivered via
Postmark but all the template rendering logic will be handled by the
Rails app.

This patch only starts with the optional `member_activated` email and
the Postmark template will still be used for now. The template edition
UI is also for now only available to me until all emails have been
migrated over to the app.

The process will be like that:
1. Support all email templates currently on Postmark
2. Update all ACP templates with the custom modification made on Postmark
3. Switch all logic to deliver emails via the app templates
4. Drop all Postmark templates
5. Make the email templates UI available to all admin/superadmin
6. Annouce the new feature! 🎉