Recurring billing, provide next billing date
This patch adds a new `#next_date` method to the `RecurringBilling`
class which provides the comming date when the recurring billing will
occured if the recurring billing is enabled.

This helps understanding how and when a specific member or membership
will be billed and makes the automatic billing feature more transparent
and less magical.