Fine-tune ActivityParticipation#reminderable?
TaPatate is senting bike lock code in the activity participation's
reminder so it is important that the email is sent even if the
participation is done at the last minute.

This patch ensures that at least always one email is sent even if the
participation is done a few days before the activity's date.