Shop, allow to close shop per delivery
This patch adds a new `shop_open` boolean to the `deliveries` table
which allows to close the shop for a specific delivery.

When the shop is closed for a delivery, the members will not be able
to open a shop order for this delivery anymore.

This developement has been requested and paid by the P2R.