Shop, add delivery order note PDF export
This patch adds a new PDF file export to generate the delivery note for
all the orders of a give shop delivery date.

This delivery note is useful to prepare all the orders and then join
it to the vegetable basket.

The delivery note is one page per shop order that includes the following
information:

- Member name
- Depot name
- Delivery date
- List of all order items with quantity and name, but without price
- Footer note that can be changed in the ACP Settings

This developement has been requested and paid by the P2R.