Shop, allow to link product to a basket complement
This patch allows to link a product to a basket complement. That way
members can order extra basket complement via shop order in addition of
the basket complement memberships.

When a product is linked to a basket complement it is allowed to only
have one single variant. So it has a one-to-one link to the basket
complement.

All product orders will later be counted in the Excel recap, the
delivery signatures PDF, and the delivery dashboard.

This developement has been requested and paid by the P2R.