Members section, switch to Hotwire (Turbo and Stimulus) with importmap
This patch updates all the javascript on the members section to use the
new importmap Rails 7 approach instead of using Webpack. On top of that
all the javascript logic has been refactored to use Stimulus and Turbo.

The refactored logic in Stimulus controller is much more simple and less
tied to the specific form which make it easier to maintain and reused.

The members stylesheets is still mainly using Webpack but that will be
changing later on and moved to TailwindCSS.