Add next basket to members index and page
This patch adds a new "next basket" column in the members index page
when the "active" scope is selected and a new "next basket" section
in the member page.

This allows to quickly see the next basket description for a member and
quickly jump to the membership page of a member.