Shop, add custom text and terms of sale URL
This patch adds a new `shop_text` and `shop_terms_of_sale_url` ACP
settings to customize the member shop page.

The shop text is displayed at the top of the member shop page and the
terms of sale URL at the bottom if they are present.

This developement has been requested and paid by the P2R.