Redesign settings page and menu
This patch redesigns the settings page by adding tabs for each section
and improving how the extra features are displayed.

The menu icons are also improved, the settings is now out of the other
submenu and has its own icon. The ACP name is also removed from the
header as it is taking up too much space for no reason.