DeliveriesCycle model introduction
This patch introduces a new model for managing delivery cycle. It will
be used to handle multiple deliveries scheme per depot, which wasn't
possible before.

The model isn't available in the UI yet, as all existing must be
migrated from the `deliveries_depots` tables to this new model.

Once the matching cycles are ready for the migration and more UI/logic
is added to manage the delivery cycle per depo, then it will be shown.