Introduce deliveries cycle model
This patch introduces a new deliveries cycle model that allows to have
different cycle of deliveries per depot. In the past duplicated depots
for the same physical location were required to support it, this is not
the case anymore. If visible, the member can pick the cycle of deliveries
when signup or renewing its membership.