Make the basket_price_extra feature available to all
This patch removes the `basket_price_extra` feature flag and makes the
feature available to all.

The way the extra is displayed on the membership page and the invoice
has also been improved to list the extra as a separate line item.

An handbook page as also been added to explain the feature configuration.