Activity, improve participations ical feed
This patch improves how the participations ical feed is displayed to
use the `webcal://` protocol instead of `https://` and disable
turbolinks so it can be directly used by the calendar application.

The participations included in the feed have also been reduced to only
includes the recent and future ones only.

The feed is also now enabled by default for all ACP.