Add permission model
This patch adds a new `Permission` model that allow to define and name
admins write access premissions.

Previoulsly, the permissions were hardcoded in the `Admin` model and it
wasn't really clear for the admins what they could or couldn't do.

Only the default "super admin" permission is mandatory and gives access
to sensible models, but all the rest can be edited (read/write).