ActivityParticipation, fix coming state logic
The `ActivityParticipation#coming?` state logic was not aligned with its
`coming` scope which depends on the `Activity.coming` scope which
consider today as coming.

This allowed a same day activity's participation to be validated or
rejected which shoudln't be possible.

The validate/reject batch actions have also been removed from the
scopes where they don't make sense.