Fix CPS with latest importmap update
The current version of es-module-shims provided
by importmap-rails (v1.4.6) requires allowing
`frame-src: blob:` in an app's CSP policy (it
creates an iframe with a blob as it's source
in order to perform feature detection). This is
quite a permissive thing to allow in csp and
I wouldn't recommend it if there is an alternative.

Since v1.4.7 es-module-shims no longer uses a
blob and it will work with `script-src: 'self'`
instead. I've upgraded to the latest version,
which is v1.5.5. For more details about the
CSP issue and it's fix see:

https://github.com/rails/importmap-rails/commit/9696db2ffc05862ffa0cee2acc56325e472caef3