Shop, handle unavailable product variant in cart
It could happen that a product or its variants is made unavailable
while it was already in the cart. In this case we need the member must
remove it before it can confirm the order.

Previously the confirm button was disabled but it wasn't clear why it
was the case, now an "unavailable" label is shown next to the order item
and the "+" button is disabled as well.