BasketContent, improve how product data is pre-filled
This patch improves how the products and unit selects are behaving when
changing not only by using the lastest basket content quantity and unit
price but by now separating then latest kg and pc data.

This ensures that a product can be used for both metrics individually
depending on the admin's choice.