Tenant interfaces and ActiveJob context improvements
This patch improves the Tenant switch interfaces by removing some of its
methods from the ACP class to reduce its boundary.

The logic to switch tenant in the ActiveJob context is also improved by
reducing its dependency to Sidekiq middleware and now support keeping
the Current class attributes as well, which will be used later on.

This should also make easier to switch to another ActiveJob adapter
later on if needed.