Aduit, track invoice and payment changes
This patch improves how the current session is tracked to facilate the
audits of the invoice and payments changes by the system or admin.

This allows to track who created or updated an payment and who created,
sent, closed, or canceled an invoice. Also if the creation of a payment
as for effect to close one or many invoices the session is tracked too.

Storing the session in `Current` also removes the need to pass it
explicitly all over the place and make it available in every context,
active job included.