Invoice, remove not_sent state
This patch removes the `not_sent` invoice state as it is confusing and
it doesn't make sense to have a dedicated state for it.

Now the invoice state machine is:
procesing => open => closed / canceled

That way an invoice can be open or close and not sent at the same time.