Shop, support special deliveries
This patch adds a new `Shop::SpecialDelivery `model that allows to open
the shop to a custom delivery outside of the usual membership delivieres.

This is useful for special sales event and it also allows to open the
shop to members that doesn't have an active membership.

The special delivery can have its own closing schedule and a custom
text description. The list of available products is also customizable.

With this feature in place, the GroupBuying will be removed once all ACP
have been migrated to the shop.

Close https://github.com/acp-admin/acp-admin/issues/55