Member account, allow depot update
This patch adds a new settings to the ACP to allow members to update
the depot of a single futur basket or all futur baskets by themselves.

All the visible depots having the membership deliveries cycle are
available to the member.

The membership price is updated accordingly.

Close https://github.com/acp-admin/acp-admin/issues/56