Remove datepicker and make filters live
This patch removes the use of the jQuery UI datepicker and replaces it
with the native HTML5 datepicker.

The filters change are also now live and refresh the index page directly
with a Stimulus controller.

The jQuery UI datepicker wasn't triggering the change event on the
date_range filter, so it was not possible to filter by date range live.

Close https://github.com/acp-admin/acp-admin/issues/59