Basket price extra, support dynamic pricing
This patch allows to define a dynamic pricing for the basket price extra
feature. This is done by defining a liquid template in the ACP settings
that will be evaluated for each membership.

This is useful for Cocagne that is using an extra price depending on the
member revenue and the basket size.