Basket complement, support deliveries cycle in memberships
This patch allows to specify a deliveries cycle for a membership basket
complement. When used, the basket complement will only be added to the
membership basket if the delivery is included in the deliveries cycle.

The basket complement own deliveries settings is still enforced.

This is useful when a member want to alternate between two or more
complements, and will be used by Cocage.

Also some new cycles results pattern have been added to support 1st,
2nd, 3rd and 4th quarter of the deliveries.