Newsletter, adds template and block system
This patch introduces the foundation for a newsletter system.

So far the only the template and block system is implemented. It allows
to create templates with blocks that can be then filled with content
when building a newsletter.

The members audience selection and the sending of the newsletter will be
implemented in a later patch.

https://github.com/acp-admin/acp-admin/issues/54