Perf, cache basket price extra
With the introduction of the price extra dynamic pricing feature, the
calculation of the price extra for a basket is quite expensive.

This patch adds a new `baskets.price_extra` cache column and updates it
when the basket is saved or the membership is updated.

This makes baskets extra price total calculation much faster.