Newsletter, unsubscribe/subscribe-back pages
This patch adds the unsubscribe link in the newsletter footer. It points
to a page using the encrypted email as a token that automatically
suppresses the member's email address from the broadcast stream.

A subscribe-back link is also added to the unsubscribe page, which is
also a page using the encrypted email as a token that unsuppresses the
member's email address from the broadcast stream.

https://github.com/acp-admin/acp-admin/issues/54