Newsletter, handle deliveries
This patch adds a new `Newsletter::Delivery` model to handle the
deliveries of a newsletter.

This allows to implement the `Newsletter#send!` method which will
persist all the audience members and send the newsletter to them one by
one.

The delivery object is also used to track the progress of the delivery,
the suppressed emails, and also keep a copy of the rendered subjet and
body of the newsletter specific to each member.

https://github.com/acp-admin/acp-admin/issues/54