Mailer, remove html tags from subject
The email template subject can be edited to include html tags. This is
not a problem for the content of the email, but it is for the subject
which is used as a header in the email.

This commit removes the html tags from the subject before sending the
email.