Newsletter, add link to subscribe back from the account page
This patch adds a link to the account page to subscribe back to the
newsletter if the user has unsubscribed from it.

If Mailchimp is configured, the link will only appear if the user has
unsubscribed from the newsletter through the website and not through
Mailchimp.

Close https://github.com/acp-admin/acp-admin/issues/77