Fix preview of newsletter and mail templates
Using the preview with a GET request was not working anymore because of
the size of the query string. The preview is now using a POST request
with the form data as body.

```
HTTP parse error, malformed request ("GET /newsletters/preview.js" - (-)): #<Puma::HttpParserError: HTTP element QUERY_STRING is longer than the (1024 * 10) allowed length (was 15027)>
```