ActionText, add image attachment support
This patch adds support for image attachments in ActionText. This allow
rich text content to be saved with images, and for those images to be
displayed either as a gallery or as a single image.

The newsletters content block, which use ActionText, now display images
as well.

All elements (shop product, extra texts, ...) that use ActionText get
the same support for images.

- https://github.com/acp-admin/acp-admin/issues/57
- https://github.com/acp-admin/acp-admin/issues/76