Newsletter, persist audience name once sent
This patch makes the newsletter audience name persistent once the
newsletter is sent. This is done by storing the audience name in the
database, and using the persisted value when the newsletter is sent.

This is done to ensure that even if a audience segment is renamed, the
newsletter will still display the name it had when it was sent.

Related to https://github.com/acp-admin/acp-admin/issues/81