Introduce member shop depot concept
This patch introduces the concept of a shop depot for members. This
allows to give access to the shop to inactive members who doesn't have a
membership. This makes those members active.

For now this is only available for ACP with the member form mode set to
"shop", but this will be extended to "membership" in a future patch.

The member subscription form has also been updated to allow to select a
shop depot when the member form mode is set to "shop".

Various updates have also been made to improve how the shop order depot
is displayed in the admin interface.

The member page now also display the list of recent shop orders.

Related to:
https://github.com/acp-admin/acp-admin/issues/90