Shop, allow to set a percentage on the order amount
This patch adds the possibility for an admin to set a percentage either
positive or negative on the order amount. This is useful for example to
give a discount to a member.

The percentage is then displayed on the invoice and order page in the
member and admin interface.

Related to:
https://github.com/acp-admin/acp-admin/issues/89