Improve member order management
This patch removes the `form_priority` attribute from the basket sizes,
basket complements, depots and deliveries cycles models and replaces it
with a `member_order_priority` attribute on top of new member_order
modes for each of these models.

Those new modes can be set on the ACP settings page and are used to
determine the order in which the corresponding models are displayed in
the member section.

Close https://github.com/acp-admin/acp-admin/issues/94