Improve delivery XLSX/PDF links and add depot delivery members list
This patch improves how the delivery XLSX/PDF links are displayed in the
dashboard, delivery, and now depot pages by using icons on the panel
header instead of a link in the panel body.

This patch also adds a new page to the depot admin section to display
the list of members for the next delivery of the depot.

The "signature sheets" have also be renamed just "sheets" as new modes
will be added in the future.