Delivery PDF sheets, add "home delivery" mode
This patch adds a new "home delivery" mode to render the delivery PDF
sheets for a depot in a way that is more suitable for home delivery.

This mode allows to sort the depot's members (via drag and drop) and
add the member delivery addresses and delivery notes to the PDF sheets.

Some various improvements have been made to the PDF sheets as well to
make it more readable overall.

Close https://github.com/acp-admin/acp-admin/issues/82