Shop, allow to set unavailable deliveries for a product
This patch allows to set unavailable deliveries for a shop product in
the admin interface. This is useful for products that will not be
available for a future delivery.

The patch also adds a filter on the products index page to filter by
deliveries and depots.

Close https://github.com/acp-admin/acp-admin/issues/78