ActivityParticipation, disable submit button when no activity available
This patch ensures that the submit button is disabled when no activity
is available for the selected date, which avoids to confuse the member
when he clicks on the button and nothing happens.

Here it was even more confusing because the success flash message was
displayed, but the activity was not created.

This patch also fixes a bug in the calendar controller, when the
selected date of another month was clicked on, the month changed but
the selected date was not kept selected.