MailTemplate, add new invoice_cancelled template
This patch adds a new invoice_cancelled mail template, which is not
active by default.

It is sent when an invoice is cancelled, but only if the invoice was
open. It is not sent when the invoice is already closed, or when the
template is not active.

Close https://github.com/acp-admin/acp-admin/issues/96