DeliveriesCycle, add minimum_gap_in_days and all_but_first results
ACP with a different deliveries schedule across they year, for example
40 deliveries but every week during summer and every 2 weeks during
winter, that want to propose a cycle every 2 weeks weren't able to do
it. As the cycle was every 2 results, it switched to every 4 weeks
during winter.

This patch introduces a new `minimum_gap_in_days` attribute that allows
to define the minimum number of days between 2 deliveries. For example
13 for a delivery every 2 weeks.

The new `all_but_first` results allows to skip the first result of the
cycle. For example, if the cycle is every 2 weeks, it will skip the
first week of the year.

Close https://github.com/acp-admin/acp-admin/issues/102