Absence, add member note field
This patch allows the member to add a note to their absence when they
declare it. This note is then displayed to the admin absences index with
a icon/tooltip.

A new `new_absence_with_note` notification is also added to allow the
admin to be notified only when the member adds a note. This is on top
of the existing `new_absence` notification which is sent when any
absence is declared.

Note, the existing `note` field on the absence model is cleared as all
its content were already moved as admin's comment in the past as it was
used only internally.

Close https://github.com/acp-admin/acp-admin/issues/104