Delivery's baskets CSV export
This patch adds a new CSV export for all baskets of a delivery, similar
to the XLSX export but in a more lightweight format.

This allows to easily process this data for example print labels for
each basket.

Some refactoring has been done to share code between the CSV and XLSX
exports and the object descriptions have been improved.

Close https://github.com/acp-admin/acp-admin/issues/107