Shop, allow to display a product in delivery sheets
This patch allows to display a product in delivery sheets (PDF, XLSX,
and CSV) if it has only one variant.

This is similar to what is already possible for product linked to a
basket complement, but it is now available for any product with a single
variant.

Close https://github.com/acp-admin/acp-admin/issues/103