Billing, add shop worksheet on the XLSX recap
This patch adds a new worksheet to the XLSX billing recap, listing all
the products sold during the year, with their quantity and total amount.

This is useful to check the total amount of products sold by category
and producer for the accounting.

Close https://github.com/acp-admin/acp-admin/issues/108