Invoice, add link to edit shop order from invoice page
As shop order's invoice cannot be directly canceled, a link to modify
the shop order (which as for effect to cancel the invoice) is added
to the invoice page. This should avoid the user to have to go back
to the shop order page to cancel the invoice or not found how to do it.