PaymentsRedistributor, stop invoice entity type prioritization
The prioritization of invoice entity types is causing more trouble than
it's worth. It makes it hard to understand how and why some payments are
distributed the way they are and it's confusing to admin and members.

This commit removes the prioritization and instead uses will try to
keep invoices that were already closed in the past closed. This should
improve the situation where already closed invoices are reopened for no
apparent reason.