Membership pricing tooltip
This patch adds a tooltip to the membership pricing to point to the last
invoice PDF in order to get more details about it.

Showing the full pricing details is complex and the whole logic is
already implemented in the invoice PDF generation. So it's easier to
just point to the PDF.