Basket price extra dynamic pricing, include complements price
This patch allows to define a dynamic price for a basket, based also on
the complements price. This is useful for example to when the basket
price is a percentage increase/discount on the basket and complements
price.

The patch also fixes a bug on the live calculation of the basket price
when the extra price is negative (discount).

As the basket price extra can now include the complements, the extra is
place after the complements in the member forms, invoices pdf, and
membership pricing description.

Finally, this patch enforces the presence of a basket price extra when
creating a member, to ensure that the extra is explicitly set by the
new member.