BasketSize, support specific deliveries cycles
This patch allows to restrict the available deliveries cycles for a
basket size. This is useful when a basket size is only available for
specific deliveries cycles, like only for the first delivery of each
month.

The deliveries cycle of a membership still needs to be available at
both its basket size and its depot.