BasketSize deliveries cycle, various fixes
This patch fixes a number of issues with the basket size delivery cycle
association to ensure that existing setup works as expected.

This patch removes support for the member new form to support basket
size selection of different delivery cycles and this is reserved for
the depot only for now.