DeliveryCycle change visibility logic
This patch simplifies the logic for determining the visibility of a
delivery cycle, which is now based on the linked depots. If at least
two cycles are available for a depot, then cycle choice is visible on
the new member form.

Depot cycles are also no more enforced when a admin create or update a
membership, this is now the responsibility of the admin to check that
the selected cycles makes sense for the selected depot. This allow more
flexibility for the admin to create or update a membership or a specific
basket delivery.

One exception is the case where a basket size is linked to a delivery
cycle, in this case the cycles choice is not visible on the new member
form.