Membership renewal fix, hide basket price extra for salary basket
This patch hides the basket price extra list in the membership renewal
form for member with the salary basket mode on. As those members aren't
paying their baskets, it's confusing to show them the price extra list
as it will not be used anyway.

Fix https://acp-admin.sentry.io/issues/4682013884