Social Share, minimum number of shares
This patch allows to define a minimum number of shares for new members,
especially the ones deciding to join as support members (shares only).

Previously, the number of shares was 1 by default. This patch allows to
define an higher number of minimum shares. It's still possible to change
that minimum number of shares for each basket size.

Close https://github.com/acp-admin/acp-admin/issues/118