Refactoring, basket state and absence_id
This patch introduces a new `state` attribute to the `Basket` model to handle
the `absent` and `trial` states in place of their boolean attributes.

A new `absence_id` attribute is also introduced to link an absent basket to its
`Absence` record. This will be handy to differentiate provisional absent
baskets.

Some Basket scopes refactoring have been done for better readability.

Related to #114