Basket, introduce billable attribute
This patch introduces a new `billable` boolean attribute to the Basket
model. This attribute is used to determine whether a basket is billable,
which is now only when the absence are configured not to be billed.

This allows an ACP to change that setting without impacting past membership
baskets.

This will also be handy to introduce included absences in the future.

Related to #114