Activity Participation, allows edit in subscription/renew forms
This patch allows to edit an min and/or max activity participations that can be
edited in a subscription or renewal form.

The default activity participations members are still updated via js when the
basket sizes or complements with the activity participations are selected, but
if the min and/or max are set in the ACP settings then the member will be able
to edit them in the subscription or renewal form.

The `activity_participations_annual_price_change` update logic has also been
improved to automatically be set by default from the diff of the default and
what is actually set in the membership but can still override if needed.

Close https://github.com/acp-admin/acp-admin/issues/49