Billing, support FR and DE IBAN format
This patch adds support for the French and German IBAN format.

Specific Swiss QR mention is also removed to improve international invoicing.

As example, the QR payment type is now named "Auto" instead of "QR".

The Invoice PDF generation still uses the Swiss QR format, but this will be
changed in a future patch.

https://github.com/acp-admin/acp-admin/issues/123